Add a test for `NetSocket#pause` and `NetSocket#resume`.
There have been intermittent failures in Nodyn on a very similar test.
This is an attempt to narrow the gap and create a failing test in vert.x
Java API. I have received intermittent errors locally, and find that
behavior changes when I change the timeout duration or buffer size.

However, I have not yet been able to produce a test that generates the
error consistently on my local development machine. I have seen the
failures primarily in continuous integration builds where the hardware
is more taxed that on a 2014 model Macbook Pro.